added quick save functionality and blocked user from quick save without already saved patch

+++ src/components/PresetsModal/PresetsModal.jsx
@@ -5,6 +5,7 @@ import { useForm } from 'react-hook-form'
 import './PresetsModal.css'
 import { ACTIONS } from '../../utils/ACTIONS';
 import toast from 'react-hot-toast'
+import { handleMouseEvent } from '../../services/general.services';
 
 const PresetsModal = ({ showSaveModal, setShowSaveModal, setCurrentPatch } ) => {
     const [ savedPatches, setSavedPatches ] = useState([])
@@ -41,32 +42,43 @@ const PresetsModal = ({ showSaveModal, setShowSaveModal, setCurrentPatch } ) => 
     const handleSaveNewPreset = async (e) => {
         const existingPresets = await JSON.parse(localStorage.getItem("savedPatches"))
         let matchingPreset
-        if (existingPresets) {
-            existingPresets.forEach(preset => {
-                if (Object.keys(preset)[0] === e.newPresetName) {
-                    setErrorMessage("Preset name already exists, please choose a different name") 
-                    matchingPreset = true 
-                }
-            })
+        if (e.newPresetName.length === 0) {
+            setErrorMessage("Preset name cannot be empty") 
+            return
         }
-       
-        if (!matchingPreset){
-            await localStorage.setItem("presetToLoad", e.newPresetName)
-            await updateState({type: ACTIONS.SYNTH.saveNewPreset, payload: {value: e.newPresetName}})
-            getSavedPatches()
-            setCurrentPatch(e.newPresetName)
-            reset()
-            setShowSaveModal(!showSaveModal)
-            toast.success(`'${e.newPresetName}' has been successfully saved`, {
-                style: {
-                    marginTop: "5rem",
-                    borderRadius: '10px',
-                    background: '#333',
-                    color: '#fff',
-                    fontFamily: "mainFont"
-                },
-              })
+        else {
+            if (existingPresets) {
+                existingPresets.forEach(preset => {
+                    if (Object.keys(preset)[0] === e.newPresetName) {
+                        setErrorMessage("Preset name already exists, please choose a different name") 
+                        matchingPreset = true 
+                    }
+                    else {
+                        setErrorMessage("") 
+                        matchingPreset = false
+                    }
+                })
+            }
+           
+            if (!matchingPreset){
+                await localStorage.setItem("presetToLoad", e.newPresetName)
+                await updateState({type: ACTIONS.SYNTH.saveNewPreset, payload: {value: e.newPresetName}})
+                getSavedPatches()
+                setCurrentPatch(e.newPresetName)
+                reset()
+                setShowSaveModal(!showSaveModal)
+                toast.success(`'${e.newPresetName}' has been successfully saved`, {
+                    style: {
+                        marginTop: "5rem",
+                        borderRadius: '10px',
+                        background: '#333',
+                        color: '#fff',
+                        fontFamily: "mainFont"
+                    },
+                  })
+            }
         }
+        
         
     }
 
@@ -77,6 +89,19 @@ const PresetsModal = ({ showSaveModal, setShowSaveModal, setCurrentPatch } ) => 
     }
 
     const handleOverwritePreset = async (e) => {
+        if (!e.overwritePreset){
+            toast.error(`No patch selected. Please select the patch you wish to overwrite`, {
+                style: {
+                    marginTop: "5rem",
+                    borderRadius: '10px',
+                    background: '#333',
+                    color: '#fff',
+                    fontFamily: "mainFont",
+                    zIndex: "20000"
+                },
+            })
+            return
+        }
         await localStorage.setItem("presetToLoad", e.overwritePreset)
         await updateState({type: ACTIONS.SYNTH.overwritePreset, payload: {value: e.overwritePreset}})
         setShowSaveModal(!showSaveModal)
@@ -105,7 +130,7 @@ const PresetsModal = ({ showSaveModal, setShowSaveModal, setCurrentPatch } ) => 
                 color: '#fff',
                 fontFamily: "mainFont"
             },
-        })
+        })  
         getSavedPatches()
     }
 
@@ -134,7 +159,7 @@ const PresetsModal = ({ showSaveModal, setShowSaveModal, setCurrentPatch } ) => 
                 <div className='new-preset-container'>
                 <label className='sliderLabel save-new-patch-label' htmlFor="saveNewPreset">Save New Patch</label>
                     <form action="saveNewPreset">
-                    {errorMessage && <p>{errorMessage}</p>}
+                    {errorMessage && <p className="preset-error-message">{errorMessage}</p>}
                         <input
                         placeholder='enter patch name' 
                         type="text" 
@@ -142,9 +167,15 @@ const PresetsModal = ({ showSaveModal, setShowSaveModal, setCurrentPatch } ) => 
                         {...register("newPresetName")}
                         />
                         <button 
-                        className='btn'
+                        className='btn save-preset-btn'
                         type="submit"
                         onClick={handleSubmit(handleSaveNewPreset)}
+                        onMouseDown={
+                            () => handleMouseEvent(`save-preset-btn`, true)
+                        }
+                        onMouseUp={
+                            () => handleMouseEvent(`save-preset-btn`, false)
+                        }
                         >
                         Save patch
                         </button>
@@ -156,9 +187,15 @@ const PresetsModal = ({ showSaveModal, setShowSaveModal, setCurrentPatch } ) => 
                     <form action="submit">
                         <PresetOptions presetSection={"overwritePreset"} registerId={register3}/>
                         <button 
-                        className='btn'
+                        className='btn overwrite-preset-btn'
                         type="submit"
                         onClick={handleOverwriteSubmit(handleOverwritePreset)}
+                        onMouseDown={
+                            () => handleMouseEvent(`overwrite-preset-btn`, true)
+                        }
+                        onMouseUp={
+                            () => handleMouseEvent(`overwrite-preset-btn`, false)
+                        }
                         >
                         Overwrite Patch
                         </button>
@@ -173,9 +210,15 @@ const PresetsModal = ({ showSaveModal, setShowSaveModal, setCurrentPatch } ) => 
                         <label className='sliderLabel save-new-patch-label' htmlFor="loadPresets">Load Patch</label>
                         <PresetOptions presetSection={"loadPreset"} registerId={register2}/>
                         <button 
-                        className='btn'
+                        className='btn load-preset-btn'
                         type="submit"
                         onClick={handleLoadSubmit(handleLoadPreset)}
+                        onMouseDown={
+                            () => handleMouseEvent(`load-preset-btn`, true)
+                        }
+                        onMouseUp={
+                            () => handleMouseEvent(`load-preset-btn`, false)
+                        }
                         >
                         Load patch
                         </button>
@@ -189,9 +232,15 @@ const PresetsModal = ({ showSaveModal, setShowSaveModal, setCurrentPatch } ) => 
                         <label className='sliderLabel save-new-patch-label' htmlFor="deletePreset">Delete Patch</label>
                         <PresetOptions presetSection={"deletePreset"} registerId={register4}/>
                         <button 
-                        className='btn'
+                        className='btn delete-preset-btn'
                         type="submit"
                         onClick={handleDeleteSubmit(handleDeletePreset)}
+                        onMouseDown={
+                            () => handleMouseEvent(`delete-preset-btn`, true)
+                        }
+                        onMouseUp={
+                            () => handleMouseEvent(`delete-preset-btn`, false)
+                        }
                         >
                         Delete patch
                         </button>
